+++ src/components/Estimates/customerEstimatesTable.jsx
@@ -9,11 +9,11 @@ import { Link, useNavigate } from "react-router-dom";
 import { Close, Edit } from "@mui/icons-material";
 import CustomIconButton from "@/components/ui-components/CustomButton";
 import { useFetchDataCustomerEstimates } from "@/utilities/ApiHooks/customer";
-import Pagination from "@/components/Pagination";
+import Pagination from "@/components/Pagination/index_dep";
 import { EstimateCategory } from "@/utilities/constants";
 import { setStateForShowerEstimate } from "@/utilities/estimates";
 import { setStateForMirrorEstimate } from "@/utilities/mirrorEstimates";
-import NewPagination from "../Pagination/new_index";
+import NewPagination from "../Pagination";
 
 const style = {
   position: "absolute",
