+++ src/components/Pagination/index.js
@@ -1,49 +1,79 @@
+import { MAX_PAGES_DISPLAYED, itemsPerPage } from "@/utilities/constants";
 import { ArrowBack, ArrowForward } from "@mui/icons-material";
-import { Box, Button } from "@mui/material";
+import { Box, Button, TextField } from "@mui/material";
 
-const Pagination = ({ totalRecords, itemsPerPage, page, setPage }) => {
-  const totalPages = Math.ceil(totalRecords / itemsPerPage);
-  const MAX_PAGES_DISPLAYED = 5;
-
-  const getPageNumbersToShow = () => {
-    if (totalPages <= MAX_PAGES_DISPLAYED) {
-      return Array.from({ length: totalPages }, (_, index) => index + 1);
-    }
-
-    const pagesToShow = [];
-    const startPage = Math.max(1, page - 2); // Display three on the first side
-    const endPage = Math.min(totalPages, startPage + MAX_PAGES_DISPLAYED - 1);
-
-    if (startPage > 1) {
-      pagesToShow.push(1);
-      if (startPage > 2) {
-        pagesToShow.push("..."); // Display ellipsis if there are skipped pages
+const NewPagination = ({ totalRecords, page, setPage,inputPage,setInputPage, isShowInput ,setIsShowInput }) => {
+    // const [page, setPage] = useState(1);
+    // const [inputPage, setInputPage] = useState("");
+    // const [isShowInput, setIsShowInput] = useState(false);
+  
+    const totalPages = Math.ceil(totalRecords / itemsPerPage);
+  
+    const getPageNumbersToShow = () => {
+      if (totalPages <= MAX_PAGES_DISPLAYED) {
+        return Array.from({ length: totalPages }, (_, index) => index + 1);
       }
-    }
-
-    for (let i = startPage; i <= endPage; i++) {
-      pagesToShow.push(i);
-    }
-
-    if (endPage < totalPages) {
-      if (endPage < totalPages - 1) {
-        pagesToShow.push("..."); // Display ellipsis if there are skipped pages
+  
+      const pagesToShow = [];
+      const startPage = Math.max(1, page - 2); // Display three on the first side
+      const endPage = Math.min(totalPages, startPage + MAX_PAGES_DISPLAYED - 1);
+  
+      if (startPage > 1) {
+        pagesToShow.push(1);
+        if (startPage > 2) {
+          pagesToShow.push("..."); // Display ellipsis if there are skipped pages
+        }
       }
-      pagesToShow.push(totalPages);
-    }
-
-    return pagesToShow;
-  };
-
-  const pageNumbersToShow = getPageNumbersToShow();
-
-  const handlePreviousPage = () => {
-    setPage((prevPage) => Math.max(prevPage - 1, 1));
-  };
-
-  const handleNextPage = () => {
-    setPage((prevPage) => Math.min(prevPage + 1, totalPages));
-  };
+  
+      for (let i = startPage; i <= endPage; i++) {
+        pagesToShow.push(i);
+      }
+  
+      if (endPage < totalPages) {
+        if (endPage < totalPages - 1) {
+          pagesToShow.push("..."); // Display ellipsis if there are skipped pages
+        }
+        pagesToShow.push(totalPages);
+      }
+  
+      return pagesToShow;
+    };
+  
+    const pageNumbersToShow = getPageNumbersToShow();
+  
+    const handlePreviousPage = () => {
+      setPage((prevPage) => Math.max(prevPage - 1, 1));
+      setIsShowInput(false);
+    };
+  
+    const handleNextPage = () => {
+      setPage((prevPage) => Math.min(prevPage + 1, totalPages));
+      setIsShowInput(false);
+    };
+  
+    const handleInputChange = (event) => {
+      setInputPage(event.target.value);
+    };
+    const HandleShowInput = () => {
+      setIsShowInput(true);
+    };
+    const HandleShowRemoveInput = () => {
+      setIsShowInput(false);
+    };
+    const handleInputKeyPress = (event) => {
+      if (event.key === "Enter") {
+        const pageNumber = parseInt(inputPage, 10);
+        if (!isNaN(pageNumber) && pageNumber >= 1 && pageNumber <= totalPages) {
+          setPage(pageNumber);
+          setInputPage("");
+          setIsShowInput(false);
+        }
+      }
+    };
+  
+    const handleInputBlur = () => {
+      setInputPage("");
+    };
   return (
     <Box
       sx={{
@@ -51,6 +81,7 @@ const Pagination = ({ totalRecords, itemsPerPage, page, setPage }) => {
         justifyContent: "space-between",
         padding: "10px",
         borderTop: "1px solid #EAECF0",
+        width: "96%",
       }}
     >
       <Button
@@ -60,7 +91,6 @@ const Pagination = ({ totalRecords, itemsPerPage, page, setPage }) => {
           borderRadius: "8px",
           textTransform: "capitalize",
           fontWeight: 500,
-          background:'white',
           ":hover": {
             border: "1px solid #D0D5DD",
             color: "#344054",
@@ -77,10 +107,14 @@ const Pagination = ({ totalRecords, itemsPerPage, page, setPage }) => {
         {pageNumbersToShow.map((pagenumber, index) => (
           <Box
             key={index}
-            onClick={() => setPage(pagenumber)}
+            onClick={() => {
+              if (typeof pagenumber === "number") {
+                setPage(pagenumber);
+                HandleShowRemoveInput();
+              }
+            }}
             sx={{
-              backgroundColor:
-              page === pagenumber ? "#8477DA" : "white",
+              backgroundColor: page === pagenumber ? "#8477DA" : "white",
               color: page === pagenumber ? "white" : "#8477DA",
               width: "40px",
               height: "40px",
@@ -88,11 +122,34 @@ const Pagination = ({ totalRecords, itemsPerPage, page, setPage }) => {
               display: "flex",
               alignItems: "center",
               justifyContent: "center",
-              cursor: "pointer",
-              border: page === pagenumber ? "1px solid #8477DA" : "1px solid #D0D5DD"
+              border: "1px solid #8477DA",
+              cursor: typeof pagenumber === "number" ? "pointer" : "default",
             }}
           >
-            {pagenumber}
+            {isShowInput && pagenumber === "..." ? (
+              <TextField
+                size="small"
+                variant="outlined"
+                type="text"
+                value={inputPage}
+                onChange={handleInputChange}
+                onKeyPress={handleInputKeyPress}
+                onBlur={handleInputBlur}
+                inputProps={{
+                  min: 1,
+                  max: totalPages,
+                  inputMode: "numeric",
+                  pattern: "[0-9]*",
+                }}
+                sx={{ width: 60 }}
+              />
+            ) : pagenumber === "..." ? (
+              <div onClick={HandleShowInput} style={{ cursor: "pointer" }}>
+                {pagenumber}{" "}
+              </div>
+            ) : (
+              pagenumber
+            )}
           </Box>
         ))}
       </Box>
@@ -103,14 +160,13 @@ const Pagination = ({ totalRecords, itemsPerPage, page, setPage }) => {
           borderRadius: "8px",
           textTransform: "capitalize",
           fontWeight: 500,
-          background:'white',
           ":hover": {
             border: "1px solid #D0D5DD",
             color: "#344054",
           },
         }}
         onClick={handleNextPage}
-        // disabled={filteredData.creatorData.length === 0}
+        disabled={page === totalPages}
       >
         Next
         <ArrowForward sx={{ color: "#344054", fontSize: 20, ml: 1 }} />
@@ -119,4 +175,4 @@ const Pagination = ({ totalRecords, itemsPerPage, page, setPage }) => {
   );
 };
 
-export default Pagination;
+export default NewPagination;
